agent(upstream): use standard QEMU on ppc64le nodes

+++ agent/testsuite-alt.sh
@@ -28,9 +28,9 @@ if ! coredumpctl_init; then
     exit 1
 fi
 
-if [[ ! -f /usr/bin/ninja ]]; then
-    ln -s /usr/bin/ninja-build /usr/bin/ninja
-fi
+[[ ! -f /usr/bin/ninja ]] && ln -s /usr/bin/ninja-build /usr/bin/ninja
+[[ ! -f /usr/bin/qemu-kvm ]] && ln -s /usr/libexec/qemu-kvm /usr/bin/qemu-kvm
+qemu-kvm --version
 
 set +e
 
@@ -57,9 +57,33 @@ echo 'int main(void) { return 77; }' > src/test/test-execute.c
 # no longer block the CI image updates.
 # See: systemd/systemd#16199
 sed -i '/def test_macsec/i\    @unittest.skip("See systemd/systemd#16199")' test/test-network/systemd-networkd-tests.py
-
 exectask "ninja-test_sanitizers_$(uname -m)" "meson test -C $BUILD_DIR --print-errorlogs --timeout-multiplier=3"
 exectask "check-meson-logs-for-sanitizer-errors" "cat $BUILD_DIR/meson-logs/testlog*.txt | check_for_sanitizer_errors"
+
+# Set timeouts for QEMU and nspawn tests to kill them in case they get stuck
+export QEMU_TIMEOUT=1200
+export NSPAWN_TIMEOUT=1200
+# Set QEMU_SMP to speed things up
+export QEMU_SMP=$(nproc)
+export SKIP_INITRD=no
+export QEMU_BIN=/usr/bin/qemu-kvm
+# Since QEMU without accel is extremely slow on the alt-arch machines, let's use
+# it only when we don't have a choice (i.e. with QEMU-only test)
+export TEST_PREFER_NSPAWN=yes
+
+## Generate a custom-tailored initrd for the integration tests
+# The host initrd contains multipath modules & services which are unused
+# in the integration tests and sometimes cause unexpected failures. Let's build
+# a custom initrd used solely by the integration tests
+#
+# Set a path to the custom initrd into the INITRD variable which is read by
+# the integration test suite "framework"
+export INITRD="/var/tmp/ci-initramfs-$(uname -r).img"
+cp -fv "/boot/initramfs-$(uname -r).img" "$INITRD"
+# Rebuild the original initrd with the dm-crypt modules and without the multipath module
+# Note: we need to built the initrd with --no-hostonly, otherwise the resulting
+#       initrd lacks certain drivers for the qemu's virt hdd
+dracut --no-hostonly --filesystems ext4 -a crypt -o multipath --rebuild "$INITRD"
 
 # As running integration tests with broken systemd can be quite time consuming
 # (usually we need to wait for the test to timeout, see $QEMU_TIMEOUT and
@@ -96,7 +120,9 @@ if [[ $NSPAWN_EC -eq 0 ]]; then
     EXECUTED_LIST=()
     INTEGRATION_TESTS=(
         test/TEST-04-JOURNAL        # systemd-journald
-        test/TEST-13-NSPAWN-SMOKE   # systemd-nspawn
+        # FIXME: This test gets stuck on C8S when calling `sysctl, possibly
+        #        related to https://bugzilla.redhat.com/show_bug.cgi?id=2098125
+        #test/TEST-13-NSPAWN-SMOKE   # systemd-nspawn
         test/TEST-15-DROPIN         # dropin logic
         test/TEST-17-UDEV           # systemd-udevd
         test/TEST-22-TMPFILES       # systemd-tmpfiles
@@ -105,7 +131,9 @@ if [[ $NSPAWN_EC -eq 0 ]]; then
         test/TEST-34-DYNAMICUSERMIGRATE
         test/TEST-45-TIMEDATE       # systemd-timedated
         test/TEST-46-HOMED          # systemd-homed
-        test/TEST-50-DISSECT        # systemd-dissect
+        # FIXME: device-mapper complains about invalid ioctl and then dies
+        #        because it can't allocate memory; needs further investigation
+        #test/TEST-50-DISSECT        # systemd-dissect
         test/TEST-54-CREDS          # credentials & stuff
         test/TEST-55-OOMD           # systemd-oomd
         test/TEST-58-REPART         # systemd-repart
@@ -126,15 +154,6 @@ if [[ $NSPAWN_EC -eq 0 ]]; then
 
         # Set the test dir to something predictable so we can refer to it later
         export TESTDIR="/var/tmp/systemd-test-${t##*/}"
-
-        # Disable nested KVM for TEST-13-NSPAWN-SMOKE, which keeps randomly
-        # failing due to time outs caused by CPU soft locks. Also, bump the
-        # QEMU timeout, since the test is much slower without KVM.
-        export TEST_NESTED_KVM=yes
-        if [[ "$t" == "test/TEST-13-NSPAWN-SMOKE" ]]; then
-            unset TEST_NESTED_KVM
-            export QEMU_TIMEOUT=1200
-        fi
 
         # Suffix the $TESTDIR of each retry with an index to tell them apart
         export MANGLE_TESTDIR=1
